Remove `is_in_database` check
This wouldn't be an issue if the user is using the lib in a single guild environment. But if in a multi-guild environment, this could "falsely" result to `True` because the `is_in_database` check is not guild specific. Removing this prevents the possibility of `md` being `None` and exception AttributeError: NoneType has no attribute `total_xp`, from being raised